Created screens

+++ Android. Module 12. Android application architecture/data/src/main/java/com/example/architecure/data/LocalDataSource.kt
@@ -6,6 +6,6 @@ import kotlinx.coroutines.flow.Flow
 interface LocalDataSource {
 
     fun getMovies(): Flow<List<Movie>>
-    fun getMovieById(movieId: Int): Flow<Movie>
+    fun getMovieById(movieId: Long): Flow<Movie>
     suspend fun refreshMovies(movies: List<Movie>)
 }
