+++ Android. Module 12. Android application architecture/data/src/main/java/com/example/architecure/data/local/dao/MovieDao.kt
@@ -19,7 +19,7 @@ interface MovieDao {
     fun getMovies(): Flow<List<Movie>>
 
     @Query("SELECT * FROM movie WHERE id = :movieId")
-    fun getMovie(movieId: Int): Flow<Movie>
+    fun getMovie(movieId: Long): Flow<Movie>
 
     @Query("DELETE FROM movie")
     suspend fun deleteMovies()
